requirements.txt
emoji==2.2.0
google-search-results==2.4.1
langchain==0.0.44
mypy==0.991
mecab-python3==0.7
numpy==1.24.0
openai==0.25.0
pandas==1.5.2
pydantic==1.10.2
requests==2.28.1
types-requests==2.28.11.7
websockets==10.4
